Add files via upload
I will continue to work as time goes by, and I keep learning.

Currently, you can decide on how long you like to perform a task and how long your breaks are going to be. The concept is based on the Pomodoro technique.

The end goal is to create a library with anything clock-related. Stopwatch, Timer, Pomodoro Technique. Challenge Hours and more.

Feel free to send me an email at fabian.rodrez@gmail.com with features requests related to the project.

Signed-off-by: rodrez <fabian.rodrez@gmail.com>